requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
yarl==1.8.2
